+++ tests/services/conftest.py
@@ -19,7 +19,6 @@ from invenio_access.permissions import any_user, authenticated_user
 from invenio_app.factory import create_api
 
 from invenio_rdm_records.proxies import current_rdm_records_service
-from invenio_rdm_records.records import RDMRecord
 
 
 @pytest.fixture(scope="module")
@@ -49,32 +48,3 @@ def authenticated_identity():
 def rdm_record_service():
     """Get the current RDM records service."""
     return current_rdm_records_service
-
-
-@pytest.fixture()
-def record_community(db, uploader, minimal_record, community, rdm_record_service):
-    """Creates a record that belongs to a community."""
-
-    class Record:
-        """Test record class."""
-
-        def create_record(self):
-            """Creates new record that belongs to the same community."""
-            # create draft
-            community_record = community._record
-            draft = rdm_record_service.create(uploader.identity, minimal_record)
-            # publish and get record
-            record = RDMRecord.publish(draft._record)
-            record.commit()
-            # TODO: remove this extra func when the `add` to a community is implemented
-            record.parent.communities.add(community_record, default=False)
-            record.parent.commit()
-            record.commit()
-            # Manually register the pid to be able to resolve it
-            record.pid.register()
-            db.session.commit()
-            rdm_record_service.indexer.index(record)
-            RDMRecord.index.refresh()
-            return record
-
-    return Record()